switch to pthread_atfork() for fork barrier handling
